% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoMatchExtended.R
\name{GetOncoTreat}
\alias{GetOncoTreat}
\title{This function plot the oncoTreat/oncoMatch result}
\usage{
GetOncoTreat(oncoTreat, nn = 20)
}
\arguments{
\item{oncoTreat}{the OncoTreat/OncoMatch output; a NES matrix}

\item{nn}{Number of drugs for each phenotype}
}
\value{
A Heatmap with p-values
}
\description{
This function plot the oncoTreat/oncoMatch result
}
