% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTopValuesMat.R
\name{GetTopValuesMat}
\alias{GetTopValuesMat}
\title{This function get top master regulators}
\usage{
GetTopValuesMat(
  mat,
  top_n = 10,
  remove.duplicates = TRUE,
  order.cluster = NULL
)
}
\arguments{
\item{top_n}{the top n values needed}

\item{order.cluster}{the clusters ordered}
}
\description{
This function get top master regulators
}
