% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoMatchExtended.R
\name{IntegrateOncoTreatNESs}
\alias{IntegrateOncoTreatNESs}
\title{This function integrate NES matrices}
\usage{
IntegrateOncoTreatNESs(nes.list, nes.weight)
}
\arguments{
\item{nes.list}{a list of nes matrices}

\item{nes.weight}{a weight vector for the matrices}
}
\value{
integrated NES
}
\description{
This function integrate NES matrices
}
\author{
Junqiang Wang
}
