% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoMatchExtended.R
\name{PlotOncoMatch}
\alias{PlotOncoMatch}
\title{This function plot the oncoTreat/oncoMatch result}
\usage{
PlotOncoMatch(
  oncoMatch,
  top.n = 50,
  break.legend = 5,
  lower.tail = FALSE,
  labs.x = "",
  labs.y = "",
  geom.text.size = 7,
  plot.all = c(FALSE, TRUE),
  return.plot = c(FALSE, TRUE),
  return.dat = c(FALSE, TRUE)
)
}
\arguments{
\item{oncoTreat}{the OncoTreat/OncoMatch output; a NES matrix}

\item{nn}{Number of drugs for each phenotype}
}
\value{
A Heatmap with p-values
}
\description{
This function plot the oncoTreat/oncoMatch result
}
