% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoMatchExtended.R
\name{PlotOncoTreat}
\alias{PlotOncoTreat}
\title{This function plot the oncoTreat/oncoMatch result}
\usage{
PlotOncoTreat(
  oncoTreat,
  top.n = 10,
  round.n = 0,
  break.legend = 10,
  geom.text.size = 4,
  remove.concentration.at.10 = TRUE
)
}
\arguments{
\item{oncoTreat}{the OncoTreat/OncoMatch output; a NES matrix}

\item{nn}{Number of drugs for each phenotype}
}
\value{
A Heatmap with p-values
}
\description{
This function plot the oncoTreat/oncoMatch result
}
