% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoMatchExtended.R
\name{proteinReliabilityScore}
\alias{proteinReliabilityScore}
\title{This function generate the protein reliability score}
\usage{
proteinReliabilityScore(vpmat)
}
\arguments{
\item{vpmat}{the vpmat function}
}
\description{
This function generate the protein reliability score
}
