% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoSignatures.R
\name{PseudoSignatures}
\alias{PseudoSignatures}
\title{This function generate the combined pseudo signature from the drug perturbed signature}
\usage{
PseudoSignatures(vpmat, method = "maximalEffects")
}
\arguments{
\item{vpmat}{The protein activity signature of the perturbed drugs}
}
\value{
The pseudo signatures of two combined drugs
}
\description{
This function generate the combined pseudo signature from the drug perturbed signature
}
\author{
Junqiang Wang
}
