% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoMatchExtended.R
\name{sigT}
\alias{sigT}
\title{This function convet the value using the sigmoid function}
\usage{
sigT(x, slope = 20, inflection = 0.5)
}
\arguments{
\item{vpmat}{the vpmat function}
}
\description{
This function convet the value using the sigmoid function
}
