% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoMatchExtended.R
\name{TxtOncoTreat}
\alias{TxtOncoTreat}
\title{This function plot the oncoTreat/oncoMatch result}
\usage{
TxtOncoTreat(oncomatch.out, nn = 10, round.n = 0)
}
\arguments{
\item{nn}{Number of drugs for each phenotype}

\item{oncoTreat}{the OncoTreat/OncoMatch output; a NES matrix}
}
\value{
A Heatmap with p-values
}
\description{
This function plot the oncoTreat/oncoMatch result
}
